pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scb-datagen"
version = "0.1.0"
description = "Automated dataset generation pipeline for the Shutdown Compliance Benchmark"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31,<3",
    "pyyaml>=6.0,<7",
    "pydantic>=2.0,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0,<9",
    "ruff>=0.4",
]

[project.scripts]
scb-generate = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py310"
line-length = 120
